refactor: PlaceReviewRating vo 패키지로 이동 (#218)
