+++ infrastructure/mysql/src/main/kotlin/kr/wooco/woocobe/mysql/placereview/PlaceReviewPersistenceMapper.kt
@@ -1,7 +1,7 @@
 package kr.wooco.woocobe.mysql.placereview
 
 import kr.wooco.woocobe.core.placereview.domain.entity.PlaceReview
-import kr.wooco.woocobe.core.placereview.domain.entity.PlaceReviewRating
+import kr.wooco.woocobe.core.placereview.domain.vo.PlaceReviewRating
 import kr.wooco.woocobe.mysql.placereview.entity.PlaceReviewImageJpaEntity
 import kr.wooco.woocobe.mysql.placereview.entity.PlaceReviewJpaEntity
